Create incognito
This code uses the selenium library to open a headless instance of the Firefox browser and navigate to a specified website. You can then interact with the website by using the methods provided by the selenium library, such as find_element and click. Once you've finished your interactions, you can close the browser using the quit method.

Please note that using a headless browser in a script may not guarantee complete anonymity, as the website may still be able to track the user's IP address and other information. Additionally, using headless browsers or scripts to automate interactions with websites is often against the terms of service of the websites and could result in account suspension or termination. As always, it's important to use technology responsibly and in accordance with the terms of service of the platforms you use.