Update incognito
This code allows the user to input their username and password, which can then be used to interact with a website in headless mode. The code uses the selenium library to launch a headless instance of the Firefox browser and navigate to a specified website, where it can perform various actions. The code also generates a random number of points earned for each search performed and calculates the total number of points earned, displaying the result along with the username.

In this code, the encrypt and decrypt functions are used to encrypt and decrypt the user's password using base64 encoding. Note that base64 encoding is not a secure encryption method, and should not be used for sensitive information in production environments. However, it provides a simple way to demonstrate how to encrypt and decrypt data in Python.

Please note that this code is for educational purposes only and should not be used for unauthorized purposes, as it may violate the terms of service of the website used and could result in account suspension or termination. Additionally, using headless browsers or scripts to automate interactions with websites may not guarantee complete anonymity, as the website may still be able to track the user's IP address and other information. As always, it's important to use technology responsibly and in accordance with the terms of service of the platforms you use.